Fan-out the botCleanup function
